but: CLI ID parser is shared and non-legacy
Move the pareser logic to a not-legacy, shared place.
